Rename README1.md to README.md
Made this the README.md, as it should be for the time being. Until the Operating System (A computer's religion) is operational. As well as the information site hosted via Squarespace, and specifically the Reality OS Wiki. There is no rush to the destination only the flow of the passage of time.